Fix an issue with warden initialization order
